Added search functinoality to overview groups table
